myClass.h, myClass.cpp Added a queue in myClass to keep the timeStamp and random number generated from randNumGenerator() Use a mutex to avoid data racing condition
main.cpp
Made two functions, one as data producer, the other as data consumer.They are called in different thread
In producer function, a number can be assigned determine how many times the randNumGenerator will be called to generate data and put the data in the queue.
In consumer function, call readLastNumAndWriteToFile to read the queue and save the data in an assigned file.
In consumer thread terminates when all the producer threads are gone and the data queue is empty.